# stateutil: Reduce allocations by specifying slice capacity eager