**What type of PR is this?**

Other

**What does this PR do? Why is it needed?**

I found that this method is historically allocating significant amounts of memory. This can cause memory churn and overallocation of the slice array.

Here are some benchmark results.

Before

```
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkUint64ListRootWithRegistryLimit
BenchmarkUint64ListRootWithRegistryLimit/100k_balances
BenchmarkUint64ListRootWithRegistryLimit/100k_balances-8                      62          17544508 ns/op        22935535 B/op     100100 allocs/op
```

After
```
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkUint64ListRootWithRegistryLimit
BenchmarkUint64ListRootWithRegistryLimit/100k_balances
BenchmarkUint64ListRootWithRegistryLimit/100k_balances-8                      78          13245190 ns/op        10737667 B/op     100070 allocs/op
```

Diff

```
name                                             old time/op    new time/op    delta
Uint64ListRootWithRegistryLimit/100k_balances-8    17.5ms ± 0%    13.2ms ± 0%   ~     (p=1.000 n=1+1)

name                                             old alloc/op   new alloc/op   delta
Uint64ListRootWithRegistryLimit/100k_balances-8    22.9MB ± 0%    10.7MB ± 0%   ~     (p=1.000 n=1+1)

name                                             old allocs/op  new allocs/op  delta
Uint64ListRootWithRegistryLimit/100k_balances-8      100k ± 0%      100k ± 0%   ~     (p=1.000 n=1+1)
```

**Which issues(s) does this PR fix?**


N/A, but should be a notacible performance improvement.

**Other notes for review**